requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
importlib-resources==6.0.1
install==1.3.5
kiwisolver==1.4.4
nptyping==2.5.0
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pynrrd==1.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tifffile==2023.8.12
typing_extensions==4.7.1
zipp==3.16.2
